sublation: Fix certificate issuance with Caddy
My previous setup had only been working temporarily; the certificates
had already been issued, so while I thought Caddy was successfully
issuing certificates, it wasn't (it had done so in a previous
attempt).  This is because the podman network it was on was internal:
internal podman networks themselves have no outbound connection
therefore, Caddy, when inside the internal network trying to make
requests to ACME services, could not do so.

Serving files over the web worked because those files were being
requested to my host on ports 80 and/or 443, with those requests being
forwarded to my Caddy container.  The problem lies with the container
network itself trying to make connections outside my host (--internal
denies that).

Therefore, we MUST use an external network for Caddy to be able to
issue certificates.